Ensure systemd-networkd-wait-online works properly (#123)
To ensure the unnumbered interfaces enter the "configured" state, which is required for `systemd-networkd-wait-online` to work properly, all dynamic address configuration mechanisms like DHCP= and IPv6AcceptRA= need to be disabled. Otherwise, the interface will remain in the "configuring" state, and systemd-networkd-wait-online will timeout for the interface.

By default, IPv6AcceptRA is enabled and controls the userspace implementation of the IPv6 RA protocol in `systemd-netword`. This means there is so no interference with FRR.